// src/app/api/admin-service/models/find-result-response-service-configuration-info.ts
/* tslint:disable */
/* eslint-disable */
import { OperationResultStatusType } from './operation-result-status-type';
import { ServiceConfigurationInfo } from './service-configuration-info';
export interface FindResultResponseServiceConfigurationInfo {
	body?: null | Array<ServiceConfigurationInfo>;
	errors?: Array<string>;
	status?: OperationResultStatusType;
	totalCount?: number;
}